refactor: package정리, Service Layer로 로직 분리
