+++ src/main/java/chess/dao/ChessDao.java
@@ -1,0 +1,16 @@
+package chess.dao;
+
+import chess.dto.ChessCellDto;
+
+import java.util.List;
+
+public interface ChessDao {
+
+    void addPosition(ChessCellDto eachPosition);
+
+    void addPositions(List<ChessCellDto> board);
+
+    List<ChessCellDto> findByGameId(String gameId);
+
+    void removePositions();
+}
